lib/, src/: Move lib/string/{strtok => strsep}/
We got rid of the last strtok(3) call, and don't want any references to
it.  It's dead.
