+++ lib/nss.c
@@ -19,7 +19,7 @@
 #include "string/strcmp/streq.h"
 #include "string/strcmp/strprefix.h"
 #include "string/strspn/stpspn.h"
-#include "string/strtok/stpsep.h"
+#include "string/strsep/stpsep.h"
 
 
 #define NSSWITCH "/etc/nsswitch.conf"
