+++ lib/port.c
@@ -23,8 +23,8 @@
 #include "string/ctype/isascii.h"
 #include "string/strcmp/streq.h"
 #include "string/strcmp/strprefix.h"
-#include "string/strtok/stpsep.h"
-#include "string/strtok/strsep2ls.h"
+#include "string/strsep/stpsep.h"
+#include "string/strsep/strsep2ls.h"
 
 
 static FILE *ports;
