+++ lib/shadow/gshadow/sgetsgent.c
@@ -16,9 +16,9 @@
 
 #include "shadow/gshadow/sgrp.h"
 #include "string/strcmp/streq.h"
-#include "string/strtok/astrsep2ls.h"
-#include "string/strtok/stpsep.h"
-#include "string/strtok/strsep2arr.h"
+#include "string/strsep/astrsep2ls.h"
+#include "string/strsep/stpsep.h"
+#include "string/strsep/strsep2arr.h"
 
 
 #if defined(SHADOWGRP) && !__has_include(<gshadow.h>)
