+++ lib/shadow/passwd/sgetpwent.c
@@ -20,8 +20,8 @@
 #include "defines.h"
 #include "prototypes.h"
 #include "string/strcmp/streq.h"
-#include "string/strtok/stpsep.h"
-#include "string/strtok/strsep2arr.h"
+#include "string/strsep/stpsep.h"
+#include "string/strsep/strsep2arr.h"
 
 
 /*
