+++ lib/shadow/shadow/sgetspent.c
@@ -23,8 +23,8 @@
 #include "prototypes.h"
 #include "sizeof.h"
 #include "string/strcmp/streq.h"
-#include "string/strtok/stpsep.h"
-#include "string/strtok/strsep2arr.h"
+#include "string/strsep/stpsep.h"
+#include "string/strsep/strsep2arr.h"
 
 
 #define	FIELDS	9
