+++ lib/string/strsep/stpsep.h
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 
-#ifndef SHADOW_INCLUDE_LIB_STRING_STRTOK_STPSEP_H_
-#define SHADOW_INCLUDE_LIB_STRING_STRTOK_STPSEP_H_
+#ifndef SHADOW_INCLUDE_LIB_STRING_STRSEP_STPSEP_H_
+#define SHADOW_INCLUDE_LIB_STRING_STRSEP_STPSEP_H_
 
 
 #include "config.h"
@@ -20,8 +20,6 @@ inline char *stpsep(char *s, const char *delim);
 // string returns-pointer separate
 // Similar to strsep(3),
 // but return the next token, and don't update the input pointer.
-// Similar to strtok(3),
-// but don't store a state, and don't skip empty fields.
 inline char *
 stpsep(char *s, const char *delim)
 {
