+++ lib/string/strsep/strsep2arr.h
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 
-#ifndef SHADOW_INCLUDE_LIB_STRING_STRTOK_STRSEP2ARR_H_
-#define SHADOW_INCLUDE_LIB_STRING_STRTOK_STRSEP2ARR_H_
+#ifndef SHADOW_INCLUDE_LIB_STRING_STRSEP_STRSEP2ARR_H_
+#define SHADOW_INCLUDE_LIB_STRING_STRSEP_STRSEP2ARR_H_
 
 
 #include "config.h"
