+++ lib/subordinateio.c
@@ -29,7 +29,7 @@
 #include "string/ctype/isascii.h"
 #include "string/sprintf/stprintf.h"
 #include "string/strcmp/streq.h"
-#include "string/strtok/strsep2arr.h"
+#include "string/strsep/strsep2arr.h"
 #include "typetraits.h"
 
 #undef NDEBUG
