+++ src/login_nopam.c
@@ -67,7 +67,7 @@
 #include "string/strcmp/strprefix.h"
 #include "string/strspn/strrspn.h"
 #include "string/strspn/stprspn.h"
-#include "string/strtok/stpsep.h"
+#include "string/strsep/stpsep.h"
 
 
  /* Path name of the access control file. */
