+++ src/suauth.c
@@ -25,7 +25,7 @@
 #include "string/strcmp/strprefix.h"
 #include "string/strspn/stpspn.h"
 #include "string/strspn/stprspn.h"
-#include "string/strtok/stpsep.h"
+#include "string/strsep/stpsep.h"
 
 
 #ifndef SUAUTHFILE
