+++ src/useradd.c
@@ -73,7 +73,7 @@
 #include "string/strcmp/strprefix.h"
 #include "string/strdup/strdup.h"
 #include "string/strspn/stpspn.h"
-#include "string/strtok/stpsep.h"
+#include "string/strsep/stpsep.h"
 #include "sysconf.h"
 
 #undef NDEBUG
